requirements.txt
numpy
opencv-python
pillow
tensorflow
scikit-learn
matplotlib